Update README.md
Hi there! I opened this project in order to create a program - a universal assistant to the collector, able to identify the collector's toy, as well as provide information about the author, year of creation, contacts of the current owner. The presence of some characteristics, such as technique, size, materials used, has greatly facilitated the lives of many collectors. I am an artist-collector, I do not know how to encode, but I hope that the need for such a scanner-analyzer program is not only mine. I hope that together we will succeed! Thank you to everyone who responds.